removed unnecessary comments, split the sidebar component into header, drawer menu and list, added logo, formatted with prettier
