+++ src/components/Footer.jsx
@@ -2,23 +2,22 @@ import React from 'react'
 import { Box } from '@mui/system'
 import { Typography } from '@material-ui/core'
 
-
-
-
 const CopyRight = () => {
-    return(
-        <Typography >{"Copyright © My Website "}{new Date().getFullYear()}</Typography>
-    )
+  return (
+    <Typography >{ "Copyright © My Website " }{ new Date().getFullYear() }</Typography>
+  )
 }
+
 export default function Footer() {
   return (
-    <Box sx={{ display: 'flex', 
-    flexDirection: 'row',
-    justifyContent: "center",
-    alignItems: "center",
-    alignContent: 'flex-end'}}>
-  <CopyRight />
-
-</Box>
+    <Box sx={ {
+      display: 'flex',
+      flexDirection: 'row',
+      justifyContent: "center",
+      alignItems: "center",
+      alignContent: 'flex-end'
+    } }>
+      <CopyRight />
+    </Box>
   )
 }
