+++ src/pages/users/Users.js
@@ -1,23 +1,16 @@
 import React from 'react'
 import { Box } from '@mui/system'
-// import { Card, Typography, CardContent, CardMedia ,CardActionArea  } from '@material-ui/core'
 import { Typography } from '@material-ui/core'
 
 export default function Users() {
   return (
     <Box
-    sx={{
+      sx={ {
         display: "flex",
         justifyContent: "center",
-        padding:"10% 10%"
-
-    }}>
-          
-     <Typography>Main Content Users</Typography> 
+        padding: "10% 10%"
+      } }>
+      <Typography>Main Content Users</Typography>
     </Box>
-    
-          
-     
-   
   )
 }
